+++ packages/abi-coder/CHANGELOG.md
@@ -1,5 +1,48 @@
 # Change Log
 
+## 0.88.1
+
+### Patch Changes
+
+- 64b90a9: feat: disable `option` length validation at `enum Option` level
+  - @fuel-ts/crypto@0.88.1
+  - @fuel-ts/errors@0.88.1
+  - @fuel-ts/hasher@0.88.1
+  - @fuel-ts/interfaces@0.88.1
+  - @fuel-ts/math@0.88.1
+  - @fuel-ts/utils@0.88.1
+
+## 0.88.0
+
+### Minor Changes
+
+- 60337b5: chore!: update `forc` to `0.59.0`
+
+### Patch Changes
+
+- 0b8e1a8: feat: find deeply nested `option` coders
+  - @fuel-ts/crypto@0.88.0
+  - @fuel-ts/errors@0.88.0
+  - @fuel-ts/hasher@0.88.0
+  - @fuel-ts/interfaces@0.88.0
+  - @fuel-ts/math@0.88.0
+  - @fuel-ts/utils@0.88.0
+
+## 0.87.0
+
+### Minor Changes
+
+- 0da455a: feat!: remove nested option size validation
+
+### Patch Changes
+
+- @fuel-ts/crypto@0.87.0
+- @fuel-ts/errors@0.87.0
+- @fuel-ts/hasher@0.87.0
+- @fuel-ts/interfaces@0.87.0
+- @fuel-ts/math@0.87.0
+- @fuel-ts/utils@0.87.0
+
 ## 0.86.0
 
 ### Minor Changes