+++ packages/merkle/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Change Log
 
+## 0.88.1
+
+### Patch Changes
+
+- @fuel-ts/hasher@0.88.1
+- @fuel-ts/math@0.88.1
+
+## 0.88.0
+
+### Patch Changes
+
+- @fuel-ts/hasher@0.88.0
+- @fuel-ts/math@0.88.0
+
+## 0.87.0
+
+### Patch Changes
+
+- @fuel-ts/hasher@0.87.0
+- @fuel-ts/math@0.87.0
+
 ## 0.86.0
 
 ### Patch Changes